requirements.txt
boto3
click
pyaml